requirements.txt
python-docx
aspose-words
